Split CallerComponent into multiple files
