+++ Dynamo20/Dynamo_UI/Dynamo20_UI.csproj
@@ -207,6 +207,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CallerComponent\Initialise.cs" />
+    <Compile Include="CallerComponent\Serialise.cs" />
+    <Compile Include="CallerComponent\Core.cs" />
     <Compile Include="Components\Adapter\CreateRequest.cs" />
     <Compile Include="Components\Adapter\CreateAdapter.cs" />
     <Compile Include="Components\Adapter\Push.cs" />
@@ -233,9 +236,9 @@
     <Compile Include="Global\Extention.cs" />
     <Compile Include="Global\GlobalSearch.cs" />
     <Compile Include="Templates\CallerValueList.cs" />
-    <Compile Include="Templates\CallerComponent.cs" />
+    <Compile Include="CallerComponent\BuildOutputAst.cs" />
     <Compile Include="Templates\ValueListView.cs" />
-    <Compile Include="Templates\CallerView.cs" />
+    <Compile Include="CallerComponent\CallerView.cs" />
     <Compile Include="Views\Adapter\CreateRequest.cs" />
     <Compile Include="Views\Adapter\CreateAdapter.cs" />
     <Compile Include="Views\Engine\Explode.cs" />
